+++ stdnet/backends/__init__.py
@@ -1,3 +1,4 @@
 from structures import * 
 from base import BackendDataServer, novalue
-from main import getdb
+from main import getdb, CacheClass
+